separated dummy example from llm example
